requirements.txt
requests
pandas
python-dotenv
psycopg2-binary
nltk
SQLAlchemy
prefect
